double for gap

+++ src/GlxDirc.cxx
@@ -70,7 +70,7 @@ int main(int argc,char** argv)
     else if ( G4String(argv[i]) == "-f" ) firstevent= atoi(argv[i+1]);
     else if ( G4String(argv[i]) == "-e" ) events    = argv[i+1];
     else if ( G4String(argv[i]) == "-l" ) lensId    = argv[i+1];
-	else if ( G4String(argv[i]) == "-wg") gap       = argv[i+1];  
+    else if ( G4String(argv[i]) == "-wg") gap       = argv[i+1];  
     else if ( G4String(argv[i]) == "-x" ) particle  = argv[i+1];
     else if ( G4String(argv[i]) == "-p" ) momentum  = argv[i+1];
     else if ( G4String(argv[i]) == "-w" ) physlist  = argv[i+1];
@@ -107,7 +107,7 @@ int main(int argc,char** argv)
   if(geometry.size()) GlxManager::Instance()->SetGeometry(atoi(geometry));
   if(radiator.size()) GlxManager::Instance()->SetRadiator(atoi(radiator));
   if(lensId.size())   GlxManager::Instance()->SetLens(atoi(lensId));
-  if(gap.size())   GlxManager::Instance()->SetGap(atoi(gap));
+  if(gap.size())   GlxManager::Instance()->SetGap(atof(gap));
   if(mcpLayout.size())GlxManager::Instance()->SetMcpLayout(atoi(mcpLayout));
   if(beamDimension.size())   GlxManager::Instance()->SetBeamDimension(atoi(beamDimension));
   if(testVal1.size())   GlxManager::Instance()->SetShift(atof(testVal1));
